backend/requirements/requirements.txt
django>=3.0,<4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary
